Updating the allplayers.js based on what is committed to production.
Removing the autodocs from the repo.

Removing the autodocs from the repo.

Removing the generated binaries from group_finder and the makefile.

Fixed some display issues with some single nodes.

Adding the ability to exclude certain treenodes from the tree.

Removing some merge conflicts.

Cache the whole url as the name of the cacheID.

[#37604937] Changing setValues to setProperties

[#37604937] Adding all dependency js files to the allplayers binary files.

[#37604937] Added a sort mechanism to each tree children.

[#37604937] Removed some of the includes from the index.html files.

Adding the docs back for easier diffs.

Adding the doc files back into the repo for an easier diff.